this is a folder for the 4everhungry app. all changes done by everyone can be checked-in here and thus tracked better.
from the previous copy that we prepared for unit3, this has the following changes
a) crash for getting search results for yelp fixed. may need to move this task
to a seperate thread using async.

b) a fix in reading image from url provided by search result.

c) using simple json package instead of default json package. not sure why the
latter is causing an issue, but for now the casting crashes are disappearing
with the simple version.

d) included a jar file for c.